requirements.txt
aiofiles==23.1.0
aiogram==3.1.0
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
black==23.9.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
fake-useragent==1.2.1
filelock==3.13.1
flake8==6.1.0
frozenlist==1.4.0
greenlet==2.0.2
humanize==4.9.0
identify==2.5.33
idna==3.4
isort==5.12.0
magic-filter==1.0.11
Mako==1.2.4
mando==0.7.1
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.6.0
psycopg2==2.9.7
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pydentic==0.0.1.dev3
pyflakes==3.1.0
python-dotenv==1.0.0
python-stdnum==1.19
pytz==2023.3
PyYAML==6.0.1
radon==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
virtualenv==20.25.0
yarl==1.9.2
